Change box-and-whiskers plots to violin plots
Two of the figures are changed to violin plots: TIGER value
distribution plot and the breakdown plot of UraLex's TIGER
values. The width of the plots in the UraLex breakdown plot
reflect the number of data points in each category whereas
the plots of the simulations and UraLex are unscaled due to
UraLex having much less plottable data points than the
simulations (which were repeated many times over).